backend/pyproject.toml
[project]
name = "course_tracker_backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.1",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary]>=3.3.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]
